requirements.txt
addict==2.4.0
matplotlib==3.7.5
numpy==1.23.4
opencv_python==4.10.0.84
PyYAML==6.0.2
tqdm==4.64.1

--extra-index-url https://download.pytorch.org/whl/cu117
torch==2.0.0
torchvision==0.15.0
--extra-index-url https://pypi.nvidia.com
